requirements.txt
flask
flask-cors
python-dotenv
groq
pymupdf